requirements.txt
streamlit
openai
langchain
tiktoken
langchain-openai
langchain_community
faiss-cpu
beautifulsoup4
